requirements.txt
asgiref==3.7.2
Django==5.0.2
django-cors-headers==4.4.0
django_csp==3.8
numpy==2.0.1
opencv-python==4.10.0.84
pillow==10.4.0
sqlparse==0.4.4
typing_extensions==4.9.0
